requirements.txt
certifi==2022.9.24
charset-normalizer==2.0.12
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.9.1
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
Telethon==1.25.2
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.12
